# openApiDocumentation.yml
...
openapi: 3.0.2
info:
  title: Challenge NodeJS LiveOn BaaS
  version: 0.1.0
  description: Bank API developed as recruitment challenge
  contact:
    name: Thiago Monteiro de Paula
    email: thmonteiropaula@gmail.com
servers:
- url: http://localhost:3000
  description: Local server
...
paths:
  "/customer":
    get:
      tags:
      - Customer
      summary: Filters a specific customer or lists all of them
      description: If you're logged in as a Customer, you're only allowed to query
        for your own data
      operationId: getCustomer
      parameters:
      - in: query
        name: cpf
        example: '14789632555'
        schema:
          type: number
        required: false
      - in: header
        name: x-access-token
        description: Access token got through [POST] /user/login
        required: true
        schema:
          $ref: "#components/headers/accessToken"
      responses:
        '200':
          description: Customer(s) listed successfully
          content:
            application/json:
              schema:
                properties:
                  customers:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/CustomerOutputObj"
                        - type: object
                          properties:
                            customerDocuments:
                              type: array
                              items:
                                allOf:
                                - $ref: "#/components/schemas/CustomerDocumentOutputObj"
                            account:
                              allOf:
                              - $ref: "#/components/schemas/AccountOutputObj"
                            user:
                              allOf:
                              - $ref: "#/components/schemas/UserInputObj"
        '401':
          description: "[UNAUTHORIZED] Missing request header 'x-access-token'"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '403':
          description: "[FORBIDDEN] One client cannot query data from another client"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '500':
          description: "[INTERNAL SERVER ERROR] Can be an error fetching customer,
            JWT authentication, or any unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
    post:
      tags:
      - Customer
      summary: Inserts a new customer on the database
      description: Only a Manager has permission to insert new Customers
      operationId: postCustomer
      parameters:
      - in: header
        name: x-access-token
        description: Access token got through [POST] /user/login
        required: true
        schema:
          $ref: "#components/headers/accessToken"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/CustomerInputObj"
              - type: object
                properties:
                  address:
                    type: array
                    items:
                      allOf:
                      - $ref: "#/components/schemas/AddressInputObj"
              required:
              - cpf
              - name
      responses:
        '201':
          description: "[CREATED] Customer inserted successfully"
          content:
            application/json:
              schema:
                properties:
                  customer:
                    allOf:
                    - $ref: "#/components/schemas/CustomerOutputObj"
        '400':
          description: "[BAD REQUEST] Missing parameters in request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '401':
          description: "[UNAUTHORIZED] Missing request header 'x-access-token'"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '403':
          description: "[FORBIDDEN] Only a manager can insert new customers"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '500':
          description: "[INTERNAL SERVER ERROR] Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
    put:
      tags:
      - Customer
      summary: Update a customer on the database
      description: Only a Manager has permission to update Customers data
      operationId: putCustomer
      parameters:
      - in: header
        name: x-access-token
        description: Access token got through [POST] /user/login
        required: true
        schema:
          $ref: "#components/headers/accessToken"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/CustomerInputObj"
              - type: object
                properties:
                  address:
                    type: array
                    items:
                      allOf:
                      - $ref: "#/components/schemas/AddressInputObj"
              required:
              - cpf
              - name
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                properties:
                  customer:
                    allOf:
                    - $ref: "#/components/schemas/CustomerOutputObj"
        '400':
          description: "[BAD REQUEST] Missing parameters in request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '401':
          description: "[UNAUTHORIZED] Missing request header 'x-access-token'"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '403':
          description: "[FORBIDDEN] Only a manager can update a customer's data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '404':
          description: "[NOT FOUND] Customer not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '500':
          description: "[INTERNAL SERVER ERROR] Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
  "/customer/address":
    patch:
      tags:
      - Customer
      summary: Inserts a new address for a customer on the database
      description: Only a Manager has permission to insert new address for Customers
      operationId: patchCustomerAddress
      parameters:
      - in: header
        name: x-access-token
        description: Access token got through [POST] /user/login
        required: true
        schema:
          $ref: "#components/headers/accessToken"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - type: object
                properties:
                  cpf:
                    type: number
                    example: 14789632555
              - type: object
                properties:
                  address:
                    allOf:
                    - $ref: "#/components/schemas/AddressOutputObj"
              required:
              - cpf
              - address
      responses:
        '201':
          description: "[CREATED] Customer address inserted successfully"
          content:
            application/json:
              schema:
                properties:
                  customer:
                    allOf:
                    - $ref: "#/components/schemas/CustomerOutputObj"
        '400':
          description: "[BAD REQUEST] Missing parameters in request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '401':
          description: "[UNAUTHORIZED] Missing request header 'x-access-token'"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '403':
          description: "[FORBIDDEN] Only a manager can insert new customers"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '404':
          description: "[NOT FOUND] Customer not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '500':
          description: "[INTERNAL SERVER ERROR] Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
    delete:
      tags:
      - Customer
      summary: Removes an address from a customer
      description: Only a Manager has permission to remove address from Customers
      operationId: deleteCustomerAddress
      parameters:
      - in: header
        name: x-access-token
        description: Access token got through [POST] /user/login
        required: true
        schema:
          $ref: "#components/headers/accessToken"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - type: object
                properties:
                  cpf:
                    type: number
                    example: 14789632555
              - type: object
                properties:
                  address:
                    allOf:
                    - $ref: "#/components/schemas/AddressOutputObj"
              required:
              - cpf
              - address
      responses:
        '200':
          description: Customer address removed successfully
          content:
            application/json:
              schema:
                properties:
                  customer:
                    allOf:
                    - $ref: "#/components/schemas/CustomerOutputObj"
        '400':
          description: "[BAD REQUEST] Missing parameters in request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '401':
          description: "[UNAUTHORIZED] Missing request header 'x-access-token'"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '403':
          description: "[FORBIDDEN] Only a manager can insert new customers"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '404':
          description: "[NOT FOUND] Customer not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
        '500':
          description: "[INTERNAL SERVER ERROR] Unexpected error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorOutput"
...
components:
...
  schemas:
    ErrorOutput:
      type: object
      properties:
        error:
          type: object
          required:
          - message
          properties:
            message:
              type: string
              description: The user message with the error reason
              example: The user message with the error reason
            details:
              type: object
              description: The error details object returned, if any. It may be a
                mongoDb object or any other
    CustomerInputObj:
      type: object
      properties:
        cpf:
          type: number
          example: 14789632555
        rg:
          type: string
          example: MG1450029
        name:
          type: string
          example: Thiago Monteiro de Paula
...
    CustomerOutputObj:
      allOf:
      - $ref: "#/components/schemas/CustomerInputObj"
      - type: object
        properties:
          _id:
            type: string
            description: The Customer's unique ID stored in the database
            example: 60664ee411751ee21019b658
          address:
            type: array
            items:
              allOf:
              - $ref: "#/components/schemas/AddressOutputObj"
    AddressInputObj:
      type: object
      properties:
        type:
          type: string
          enum:
          - HOME
          - BUSINESS
          - BILLING
          - OTHER
          example: HOME
        streetName:
          type: string
          example: Nameless st
        streetNumber:
          type: number
          example: 123
        city:
          type: string
          example: Franca
        state:
          type: string
          example: SP
        postalCode:
          type: string
          example: '14400000'
    AddressOutputObj:
      allOf:
      - $ref: "#/components/schemas/AddressInputObj"
      - type: object
        properties:
          _id:
            type: string
            description: Unique ID of this address stored on database
            example: 60664ee411751ee21019b659
    AccountInputObj:
      type: object
      properties:
        agency:
          type: number
          example: 1234
        accountNumber:
          type: number
          example: 1001287
    AccountOutputObj:
      allOf:
      - $ref: "#/components/schemas/AccountInputObj"
      - type: object
        properties:
          _id:
            type: string
            description: The account's unique ID stored in the database
            example: 60665a82beda3670757f2c4b
          balance:
            type: number
            description: The account's current balance value
            example: 1200
          overdraft:
            type: number
            description: The credit value allowed for this client to stay with negative
              balance
            example: 1000
...
    CustomerDocumentInputObj:
      type: object
      properties:
        documentName:
          type: string
          description: The name is unique for each customer and may contain spaces
          example: Birth certificate
        documentType:
          type: string
          description: The document file MIME type
          example: application/pdf
    CustomerDocumentOutputObj:
      allOf:
      - $ref: "#/components/schemas/CustomerDocumentInputObj"
      - type: object
        properties:
          _id:
            type: string
            description: The document's unique ID stored in the database
            example: 6065502b7145a080110b6d50
    UserInputObj:
      type: object
      properties:
        profile:
          type: string
          enum:
          - MANAGER
          - CLIENT
          example: MANAGER
        userName:
          type: string
          description: The user name that will be used to log in the API
          example: thmonteiropaula
...